fabric_gen: Generate configMem files only if needed
Add check if configMem files needed to be generated for a tile.
Tile_configMem.csv and also Tile_configMem.v were always generated, even if they are not needed for a tile.

Remove hardcoded NoConfigBit parameters in eFPGA.v and supertile tops, since they are not needed and cause linter issues.

Signed-off-by: Jonas K. <j0nask@proton.me>