fix!: Make VHDL attribute handling and belMap processing more robust and fix broken BELs
Update Yosys_obj to keep BelMap dicts in sync with legacy bel parsing and updated
yosys json parsing.

Refactor belMapProcessing and remove hardcoded handling of INIT vector.
Allowing now defining BelMaps for all parameters and vector.
Use the actual bit mapping value of the BelMaps to define the mapping of
the config bits, instead of relying on the order of the belMap
definition.

Add checks for right order of BelMap config bit matching.

BREAKING: The lowest value of a belMap vector is now defined as
<parameter>[0] which was previously just <parameter>
This will make the bitstream spec and bel.v2.txt more consistent.

Add a "FAB_ATTR_" prefix for BelMap values, which gets removed in
the further internal processing.
This allows to define attributes with names,
which usually reserved for internal use of the synthesis tool.
Like the "FF" attribute, which is internally reserved for GHDL.

Fix errors in belMaps of:
LUT4c_frame_config_dffesr.vhdl
Config_access.vhdl
OutPass4_frame_config_mux.vhdl

Signed-off-by: Jonas K. <j0nask@proton.me>