fix!: Leave the lowest bit of a belMap vector untouched
Commit 483b3e6ca6fabfce9fa3fe1def9fe3611d6444dd
introduced a change in the belMap processing, which changed the
lowest value of a belMap vector from <parameter> to <parameter>[0].
Sadly, nextpnr has for some cases still the <parameter> notation
hardcoded, and ignores the names given in the bel.v2.txt.

This patch rolls back to the initial notation, jut <parameter> for
the lowest bit of a belMap vector.

Signed-off-by: Jonas K. <j0nask@proton.me>