chore: Write models_pack path as relative in .env (#546)
Write the models_pack path als relative path in the project .env to allow the
FABuluos project to be moved, without an error of models_pack not found.
This is already supported in the models_pack field validator and need no
further changes. The field validator still publishes the path as
absolute path for internal usage.

Add a log message with the models pack path.

Signed-off-by: Jonas K. <j0nask@proton.me>